backend/microsservicos/usuario/requirements.txt
fastapi
uvicorn
ffmpeg-python
pika
pydantic
pytest
sqlalchemy
passlib[bcrypt]
pymysql
requests
PyJWT==2.6.0
pytest-mock
pydantic